multi-paned: start on a multi-paned implementation
The goal is to be sort of like GtkPaned, but have multiple resize grips
to allow tweaking each child.

We are definitely trying to take a simplification to the problem here by
using fractions to track individual sizes. This will ultimiately have
less flexibility in terms of potential use cases, but simplify our case
rather nicely. Contrast this with the shrink/resize child properties of
GtkPaned and how it reacts to widget resizing.