overlay: make PnlDockOverlay an event box
By making this an event box, we get a GdkWindow to contain our overlay.
This keeps drawing in the panel edges from overflowing the handles of
the parent window (messing up drawing optimizations and stealing our
resize handles).

We might need to follow up a bit on the child properties, but this is
suitable for testing.